% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoival.R
\docType{class}
\name{nanoival-class}
\alias{nanoival-class}
\alias{nanoival}
\alias{start,nanoival-method}
\alias{end,nanoival-method}
\alias{sopen,nanoival-method}
\alias{eopen,nanoival-method}
\alias{print,nanoival-method}
\alias{show,nanoival-method}
\alias{names,nanoival-method}
\alias{names<-,nanoival-method}
\alias{as.nanoival,character-method}
\alias{<,nanoival,nanoival-method}
\alias{<=,nanoival,nanoival-method}
\alias{>,nanoival,nanoival-method}
\alias{>=,nanoival,nanoival-method}
\alias{==,nanoival,nanoival-method}
\alias{!=,nanoival,nanoival-method}
\alias{-,nanoival,ANY-method}
\alias{-,nanoival,integer64-method}
\alias{-,nanoival,numeric-method}
\alias{-,ANY,nanoival-method}
\alias{+,nanoival,ANY-method}
\alias{+,nanoival,integer64-method}
\alias{+,nanoival,numeric-method}
\alias{+,ANY,nanoival-method}
\alias{+,integer64,nanoival-method}
\alias{+,numeric,nanoival-method}
\alias{+,nanoival,nanoival-method}
\alias{Arith,nanoival,ANY-method}
\alias{Compare,nanoival,ANY-method}
\alias{Logic,nanoival,ANY-method}
\alias{Logic,ANY,nanoival-method}
\alias{Math,nanoival-method}
\alias{Math2,nanoival-method}
\alias{Summary,nanoival-method}
\alias{min,nanoival-method}
\alias{max,nanoival-method}
\alias{range,nanoival-method}
\alias{Complex,nanoival-method}
\alias{[,nanoival,logical-method}
\alias{[,nanoival,numeric-method}
\alias{[<-,nanoival,logical,ANY,nanoival-method}
\alias{c.nanoival}
\alias{t,nanoival-method}
\alias{intersect,nanoival,nanoival-method}
\alias{union,nanoival,nanoival-method}
\alias{setdiff,nanoival,nanoival-method}
\alias{[,nanotime,nanoival-method}
\alias{intersect,nanotime,nanoival-method}
\alias{setdiff,nanotime,nanoival-method}
\alias{is.unsorted,nanoival-method}
\alias{sort,nanoival-method}
\alias{seq,nanotime-method}
\title{Interval type with nanosecond precision}
\usage{
nanoival(start, end, sopen = FALSE, eopen = TRUE)

\S4method{start}{nanoival}(x)

\S4method{end}{nanoival}(x)

\S4method{sopen}{nanoival}(x)

\S4method{eopen}{nanoival}(x)

\S4method{print}{nanoival}(x, ...)

\S4method{show}{nanoival}(object)

\S4method{names}{nanoival}(x)

\S4method{names}{nanoival}(x) <- value

\S4method{as.nanoival}{character}(x, format = "", tz = "")

\S4method{<}{nanoival,nanoival}(e1, e2)

\S4method{<=}{nanoival,nanoival}(e1, e2)

\S4method{>}{nanoival,nanoival}(e1, e2)

\S4method{>=}{nanoival,nanoival}(e1, e2)

\S4method{==}{nanoival,nanoival}(e1, e2)

\S4method{!=}{nanoival,nanoival}(e1, e2)

\S4method{-}{nanoival,ANY}(e1, e2)

\S4method{-}{nanoival,integer64}(e1, e2)

\S4method{-}{nanoival,numeric}(e1, e2)

\S4method{-}{ANY,nanoival}(e1, e2)

\S4method{+}{nanoival,ANY}(e1, e2)

\S4method{+}{nanoival,integer64}(e1, e2)

\S4method{+}{nanoival,numeric}(e1, e2)

\S4method{+}{ANY,nanoival}(e1, e2)

\S4method{+}{integer64,nanoival}(e1, e2)

\S4method{+}{numeric,nanoival}(e1, e2)

\S4method{+}{nanoival,nanoival}(e1, e2)

\S4method{Arith}{nanoival,ANY}(e1, e2)

\S4method{Compare}{nanoival,ANY}(e1, e2)

\S4method{Logic}{nanoival,ANY}(e1, e2)

\S4method{Logic}{ANY,nanoival}(e1, e2)

\S4method{Math}{nanoival}(x)

\S4method{Math2}{nanoival}(x, digits)

\S4method{Summary}{nanoival}(x, ..., na.rm = FALSE)

\S4method{min}{nanoival}(x, ..., na.rm = FALSE)

\S4method{max}{nanoival}(x, ..., na.rm = FALSE)

\S4method{range}{nanoival}(x, ..., na.rm = FALSE)

\S4method{Complex}{nanoival}(z)

\S4method{[}{nanoival,logical}(x, i, j, ..., drop = FALSE)

\S4method{[}{nanoival,numeric}(x, i, j, ..., drop = FALSE)

\S4method{[}{nanoival,logical,ANY,nanoival}(x, i, j, ...) <- value

\method{c}{nanoival}(...)

\S4method{t}{nanoival}(x)

\S4method{intersect}{nanoival,nanoival}(x, y)

\S4method{union}{nanoival,nanoival}(x, y)

\S4method{setdiff}{nanoival,nanoival}(x, y)

\S4method{[}{nanotime,nanoival}(x, i, j, ..., drop = TRUE)

\S4method{intersect}{nanotime,nanoival}(x, y)

\S4method{setdiff}{nanotime,nanoival}(x, y)

\S4method{is.unsorted}{nanoival}(x, na.rm = FALSE, strictly = FALSE)

\S4method{sort}{nanoival}(x, decreasing = FALSE, ...)

\S4method{seq}{nanotime}(from, to = NULL, by = NULL, length.out = NULL,
  along.with = NULL, ...)
}
\arguments{
\item{x}{The object which want to convert to class \code{nanotime}}

\item{...}{further arguments passed to or from methods.}

\item{object}{argument for method \code{show}}

\item{value}{argument for \code{nanotime-class}}

\item{tz}{Required for \code{as.POSIXct} and \code{as.POSIXlt},
can be set via \code{options("nanotimeFormat")} and uses \sQuote{UTC} as
a default and fallback}

\item{e1}{Operand of class \code{nanoival}}

\item{e2}{Operand of class \code{nanoival}}

\item{digits}{Required for \code{Math2} signature but ignored here}

\item{na.rm}{a logical indicating whether missing values should be removed.}

\item{z}{Required for \code{Complex} signature but ignored here}

\item{i}{index specifying elements to extract or replace.}

\item{j}{Required for \code{[} signature but ignored here}

\item{drop}{Required for \code{[} signature but ignored here}

\item{recursive}{argument for method \code{c}}
}
\value{
A nanoival object
}
\description{
\code{nanoival} is a time interval type (an S4 class) with
nanosecond functionality. One of its purposes is to allow quick
subsetting of a \code{nanotime} vector. \code{nanoival} is
composed of a \code{nanotime} pair which defines the start and end
of the time interval. Additionally, it has a pair of logical
values which determine if the start and end of the time interval
are open (true) or closed (false).
}
\details{
An interval object can be constructed with the constructor
\code{nanoival} which takes as arguments two \code{nanotime}
objects that define the start and the end of the interval,
together with two \code{logical} arguments that define if the
start and the end of the interval are open (true) or closed
(false) (note that these objects can all be vector, and therefore
the interval object is not necessarily scalar). Alternatively, an
interval can be constructed with a \code{character}: the format
follows that of \code{nanotime}; the start time is preceeded by
either \code{-} or \code{+} indicating if the interval start is
open (-) or closed (+); the start and end times are separated by
an arrow \code{->}; the end is folloed by either \code{-} or
\code{+} which have the same semantics as the start time.

The most important set of methods defined for \code{interval} are
set functions \code{intersect}, \code{union} and \code{setdiff}.

Additionally, \code{interval} allows the subsetting into a
\code{nanotime} vector. Note that subsetting is allowed only if
the \code{nanotime} vector is sorted.

Finally, accessors are provided to get the interval start
(\code{start}), the end (\code{end}), the open/close status of the
start (\code{sopen}) and the open/close status of the end
(\code{eopen}). The former return a \code{nanotime} while the
latter return a \code{logical}.
}
\section{Output Format}{


Formatting and character conversion for \code{nanoival} objects is
identical to \code{nanotime} objects. The default format is
ISO3339 compliant: \code{\%Y-\%m-\%dT\%H:\%M:\%E9S\%Ez}. It
specifies a standard ISO 8601 part for date and time --- as well
as nine digits of precision for fractional seconds (down to
nanoseconds) and on offset (typically zero as we default to UTC).
It can be overriden by using \code{options()} with the key of
\code{nanotimeFormat} and a suitable value. Similarly,
\code{nanotimeTz} can be used to select a different timezone.
}

\examples{
x <- as.nanoival("-2012-03-01T21:21:00.000000001+00:00->2015-01-01T21:22:00.000000999+04:00+")
print(x)
}
\author{
Leonardo Silvestri
}
